+++ src/browser/js/app/widgets/sliders/slider.js
@@ -99,14 +99,6 @@ class Slider extends Canvas {
                 this.showValue()
             })
 
-            this.widget.addEventListener('fast-right-click',function(e){
-                if (!EDITING) {
-                    e.stopPropagation()
-                    e.preventDefault()
-                    this.input.canvas.focus()
-                }
-            }.bind(this))
-
         }
 
         this.setValue(this.springValue)